Create dummy OpenGL object for proper context.
As mentioned in the Qt6 OpenGL documentation, the behavior has changed
such that the first time an OpenGL widget is added to the hierarchy,
the entire top level window is flushed.  This makes it look like the
entire application restarts when the user opens an item which utilizes
OpenGL.  This change causes it to happen at the time of application
start-up instead by adding an empty hidden QtOpenGL widget to the
window.  I was unable to find any other method in the documentation to
accomplish this, and the widget does need to survive the constructor
in order to have the desired effect.